Update wordle_blocker.plugin.js
Ver 1.18

Fixed DOM'ing issues with users trying to break it, thx friends on a few different discords for helping me test. This should let people use the word Wordle and a green square, and get thru, while still blocking the wordle auto-post.

Should also not crash or close your main chat window when it triggers a run.

Stable: Tested for a few hours, no wordle scores in a heavily populated channel that posts them frequently, #main on PaymoneyWubby discord. Also tested in spam-warehouse with friends. OPEN ISSUE IF THERE'S AN ISSUE AND YOU NEED A TISSUE.